backend/requirements.txt
fastapi==0.115.0
mangum==0.17.0
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.9.1
boto3==1.35.36
Pillow==11.0.0
uvicorn==0.32.0
python-multipart==0.0.12
